SM_SO,SM_SODao  - Add campo de token
 SM_SO_Service_Exec , Task e File Dao
  - Corrigido problema nos metodos addIUpdateTmp dos Dao's, pois estavam usando a constante TMP , mas o get da propriedade code

DatabaseHelperMult
   - Add Campo token na tabela SM_SO

TSO_Serial_Save_Env
   - Add "marcador" Expose em todas a propriedade, para que os dados fossem enviados. Mudança necessaria por para enviar apenas o dados necessario da SO utilizando o metodo  excludeFieldsWithoutExposeAnnotation do Gson no pacote de envio.

SM_SO_Service_Exec_Task_Sql_003
   - Corrigido o tipo da variavel exec_tmp

SM_SO_Sql_005
   - Query que seleciona todas as S.O's com flag uodate_required marcada como 1

Act021
   - Criado metodo teste para chamar o WS que salva os dados do S.O.

+++ app/src/main/java/com/namoadigital/prj001/sql/SM_SO_Service_Exec_Task_Sql_003.java
@@ -22,9 +22,9 @@ public class SM_SO_Service_Exec_Task_Sql_003 implements Specification {
     private int category_price_code;
     private int service_code;
     private int service_seq;
-    private int exec_tmp;
+    private long exec_tmp;
 
-    public SM_SO_Service_Exec_Task_Sql_003(long customer_code, int so_prefix, int so_code, int price_list_code, int pack_code, int pack_seq, int category_price_code, int service_code, int service_seq, int exec_tmp) {
+    public SM_SO_Service_Exec_Task_Sql_003(long customer_code, int so_prefix, int so_code, int price_list_code, int pack_code, int pack_seq, int category_price_code, int service_code, int service_seq, long exec_tmp) {
         this.customer_code = customer_code;
         this.so_prefix = so_prefix;
         this.so_code = so_code;
